refactor: enhance error handling and authorization checks across order, product, and shop actions; add viewport height setter component

+++ src/actions/product/product-actions.ts
@@ -1,11 +1,13 @@
 "use server";
 
-import { Product } from "@prisma/client";
+import { revalidatePath } from "next/cache";
 
-import { InternalServerError } from "@/lib/custom-error";
+import { InternalServerError, UnauthorizedError } from "@/lib/custom-error";
 import { serializeProduct } from "@/lib/utils-functions";
 import authUtils from "@/lib/utils-functions/auth.utils";
+import { shopRepository } from "@/repositories";
 import productRepository from "@/repositories/product.repository";
+import { SerializedProduct } from "@/types/product.types";
 import { ActionResponse, createSuccessResponse } from "@/types/response.types";
 import {
   ProductActionFormData,
@@ -15,9 +17,17 @@ import {
 
 export async function createProductAction(
   formData: ProductActionFormData
-): Promise<ActionResponse<Product>> {
+): Promise<ActionResponse<SerializedProduct>> {
   try {
-    const shop_id = await authUtils.getShopId();
+    const user_id = await authUtils.getUserId();
+    const context = await shopRepository.findByOwnerId(user_id, {
+      select: { id: true },
+    });
+    if (!context || !context.id) {
+      throw new UnauthorizedError("User is not authorized to create a product");
+    }
+
+    const shop_id = context.id;
 
     const parsedData = productActionSchema.parse(formData);
     const newProduct = await productRepository.create({
@@ -27,9 +37,14 @@ export async function createProductAction(
       },
     });
 
-    // TODO: revalidate
+    const serializedProduct = serializeProduct(newProduct);
 
-    return createSuccessResponse(newProduct, "Product created successfully");
+    revalidatePath(`/shop/${shop_id}`);
+
+    return createSuccessResponse(
+      serializedProduct,
+      "Product created successfully"
+    );
   } catch (error) {
     console.error("CREATE PRODUCT ERROR:", error);
     throw new InternalServerError("Failed to create product.");
@@ -44,17 +59,23 @@ interface UpdateProductActionFormData
 export async function updateProductAction(
   product_id: string,
   formData: UpdateProductActionFormData
-) {
+): Promise<ActionResponse<SerializedProduct>> {
   try {
+    const user_id = await authUtils.getUserId();
+    const context = await shopRepository.findByOwnerId(user_id, {
+      select: { id: true },
+    });
+    if (!context || !context.id) {
+      throw new UnauthorizedError("User is not authorized to create a product");
+    }
     const parsedData = productActionSchema.parse(formData);
 
-    console.log("Updating product with data:", parsedData);
     const updatedProduct = await productRepository.update(product_id, {
       data: parsedData,
     });
     const serializedProduct = serializeProduct(updatedProduct);
 
-    // TODO: revalidate
+    revalidatePath(`/shop/${updatedProduct.shop_id}`);
 
     return createSuccessResponse(
       serializedProduct,
@@ -66,8 +87,18 @@ export async function updateProductAction(
   }
 }
 
-export async function deleteProductAction(product_id: string) {
+export async function deleteProductAction(
+  product_id: string
+): Promise<ActionResponse<null>> {
   try {
+    const user_id = await authUtils.getUserId();
+    const context = await shopRepository.findByOwnerId(user_id, {
+      select: { id: true },
+    });
+    if (!context || !context.id) {
+      throw new UnauthorizedError("User is not authorized to create a product");
+    }
+
     await productRepository.delete(product_id);
     return createSuccessResponse(null, "Product deleted successfully");
   } catch (error) {
